requirements.txt
discord.py==2.4.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.45
httpx==0.27.0
python-dotenv==1.0.1
Jinja2==3.1.4
itsdangerous==2.2.0
python-multipart==0.0.12
